backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
python-multipart==0.0.9

# LangChain + LangGraph + LangSmith
langchain==0.2.0
langgraph==0.2.0
langchain-groq==0.1.9
langchain-community==0.2.0
langsmith==0.1.0

# Video processing
ffmpeg-python==0.2.0
opencv-python-headless==4.10.0.82
imagehash==4.3.1
Pillow==10.4.0

# Audio
openai-whisper==20231117
openai==1.35.0

# OCR
easyocr==1.7.1

# HTTP
httpx==0.27.0
aiofiles==23.2.1

# Dev
pytest==8.2.0
pytest-asyncio==0.23.0
ruff==0.5.0
mypy==1.10.0
